Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -5,6 +5,6 @@ int main(){
     int a, b, c;
     scanf(%d %d %d,&a, &b, &c);
     int d = a + b + c;
-    int e = d / 3;
-    printf("Average: %d", e);
+    float e = d / 3;
+    printf("Average: %.2f", e);
 }
